--- a/octoprint_PrintJobHistory/templates/PrintJobHistory_tab.jinja2
+++ b/octoprint_PrintJobHistory/templates/PrintJobHistory_tab.jinja2
@@ -306,8 +306,11 @@
     </div>
     <!-- END: PAGEINATION BLOCK    -->
 
+    <div data-bind="visible: isLoadingTablePage" style="margin: 16px; text-align: center;">
+        <i class='fa fa-2xl fa-spinner fa-spin' ></i>
+    </div>
 
-    <table id="printjobhistory_main_table" class="table table-striped table-hover table-condensed table-hover" style="clear: both;">
+    <table id="printjobhistory_main_table" class="table table-striped table-hover table-condensed table-hover" style="clear: both;" data-bind="visible: isNotLoadingTablePage">
         <thead>
             <tr>
 <!--                <th style="width: 4%; text-align: center;" ><input type="checkbox" data-bind="checked:$root.printJobHistoryTableHelper.allSelected"></th>-->