Closes #3 

### Changelog
- [x] Introduce jobs table lazy loading
- [x] Add table page loading spinner
- [ ] Make lazy loading toggleable

### Benchmarks

#### Octoprint startup time

- Until everything is loaded
- Test repeated 5 times after initial hard reload & stabilisation
- Test performed on OrangePi Zero 2, Armbian build 2023-06-21

| - | Before | After | Δ |
|--------|--------|--------|--------|
| Load time (avg) | 5,766 s | 5,38 s | -0,386 s |